+++ core/include/pkx/PK6.hpp
@@ -245,7 +245,7 @@ public:
     inline u8 genderType(void) const override { return PersonalXYORAS::gender(formSpecies()); }
     inline u8 baseFriendship(void) const override { return PersonalXYORAS::baseFriendship(formSpecies()); }
     inline u8 expType(void) const override { return PersonalXYORAS::expType(formSpecies()); }
-    inline u8 abilities(u8 n) const override { return PersonalXYORAS::ability(formSpecies(), n); }
+    inline u16 abilities(u8 n) const override { return PersonalXYORAS::ability(formSpecies(), n); }
     inline u16 formStatIndex(void) const override { return PersonalXYORAS::formStatIndex(formSpecies()); }
 };
